+++ NEWS.md
@@ -1,3 +1,8 @@
+1.x.x [201X-XX-XX]
+==================
+
+* Improve file progress callback code.
+
 1.2.0 [2017-02-19]
 ==================